Add port-finding helper to clone box
This tries to help users figure out which port they need by asking
them to unplug, and then re-plug their cable. For USB devices, we
should be able to see the port leave and return, and if that is the
only thing that changed, determine that must be the port they want.
It won't work for fixed serial devices (of course), but most people
are using USB now.

Tested on windows and macos and was surprised at how quickly I could
click the buttons after plugging and unplugging and have it still
work.